refactor(android): decompose AndroidPeripheral 380->236 via extension functions (#306)
Extract GATT operations to AndroidPeripheralGattOps.kt (refresh,
read/write, observe, descriptor ops, RSSI, MTU) and PHY/connection
parameter operations to AndroidPeripheralConnectionParams.kt.

Second-pass decomposition following the same Approach B (extension
functions in package) used in #244.  Facade now 236 lines (target <300).
